Do not use sys.exit in library
Raise an exception in case of error instead. sys.exit will exit Python, which
is not desireable when running as a library.  Return RC 1 in these cases when
running as a script.

Move the chdir into a finally block, to ensure it is always run.  Reraise any
exceptions caught to preserve traceback.