Add a test for Unicode normalization in filenames
This test confirms that changes to the Unicode normalization form stored
on the filesystem will not cause bag validation to fail. This is
necessary to avoid failures when e.g. a bag which has filenames
normalized as a form other than NFD is transfered to a Mac HFS+
filesystem or when an intermediary tool performs Unicode normalization
unexpectedly.

With this commit, the test will reliably fail. If the normalization
forms are swapped the test will fail on Linux and Windows but pass when
run on macOS.