Add some hints on invoking the script using `-m`
Working with some baggers today, I discovered that on both Windows and Mac, it was non-obvious what it means to run the command-line program. On the other hand, if they've run `pip`, they likely can also run `python` or `py` (Windows). This change provides a hint on how to do just that.