Better visibility for duplicate file entries (see #94)
Now bagit.py will issue a warning log message if a filename
is listed twice in the same manifest with the same value.

If a filename is listed twice with different hash values it
will raise an exception with a very informative message to
make it easier for the user to identify the problem than the
standard bag validation error.