follow_links
This commit adds a follow_links parameter to Bag.make_bag and
Bag.__init__. When set to true Bag._is_dangerous will allow
links to files that exist outside of the payload directory. By
default it is set to False, but I think a reasonable argument
could be made for making the default set to True.

See #115 for more context.