follow_links and os.walk
There were several places in the code where filesystem walking happens.
These needed to be adjusted to take the follow_links option that
is provided to make_bag or the Bag constructor. In addition there are a
few functions outside of the Bag class which walk the fileystem which
needed to be aware of whether the user wants to follow links.

Interestingly this work surfaced a bit of a bug in bagit-python. If the
directory to be bagged contains a symlink before bagging, bagit will happily
atomically move that directory into the bag's payload directory. However
when the manifests are created they totally ignore the symbolic links.
So the data is there present on the filesystem but totally not
represented in the manifests.

With the addition of --follow-links you can now create a bag that
contains symlinks, and verify it later. I think an argument could be
made that this should be the default behavior, and perhaps you should
only be able to turn it off?