Fix edge case in wrapping tag values
If the given line length is shorter than the length of a tag/property
name, then we should start the value on the next line in order to try
to keep to the desired length.  This commit adds a test for that
situation and makes line wrapping behave a little bit better.