Use pyproject and remove setup.py
This commit includes changes to remove the use of setup.py and
setuptools in favor of using the existing pyrproject.toml and hatch.

A source layout was chosen, which required several changes such as:

- pushing bagit.py into src/bagit/
- pushing locale into src/bagit/locale

The gettext related code in setup.py was relocated to utils/locales.py
It's possible the way that the gettext functionality is setup needs to
adjustment, as I'm new to using it.

Also we now rely on running tests with uv and pytest, which happens in the test
github action. We could chose to update the PyPI publishing step to use
uv build/publish too?