Avoid unnecessary file checking operations
Combines _can_bag and _can_read into one function. Instead of compiling
a list of every single file and directory that cannot be read/written
to, this new function throws an Exception as soon as a problematic file
or directory comes up.

At most, one walk() operation is completed now, instead of two.